requirements.txt
# ============================================================================
# FILE: requirements.txt
# ============================================================================

streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.1
langchain-classic>=0.1.0  # For backward compatibility with AgentExecutor and create_tool_calling_agent
astrapy>=0.7.0
python-dotenv>=1.0.0